refactor(ui): move common/time_range to dbflux_components (S5) (#102)
